+++ src/error.rs
@@ -13,3 +13,25 @@ impl fmt::Display for TryFromSpellError {
         write!(f, "No value matching `{}` for attribute `{}`", self.value, self.defindex)
     }
 }
+
+impl std::error::Error for TryFromSpellError {}
+
+/// An error when appending to a set.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum InsertError {
+    /// The set is full.
+    Full,
+    /// The item is already in the set.
+    Duplicate,
+}
+
+impl fmt::Display for InsertError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            InsertError::Full => write!(f, "Set is full."),
+            InsertError::Duplicate => write!(f, "Item already exists."),
+        }
+    }
+}
+
+impl std::error::Error for InsertError {}